Create startSession
Each start(Focus, ShortBreak, LongBreak) methods implemented its own code, but since they are all similar, I created a more abstract and flexible startSession method, that each one of the start methods calls.

Now allows to call startSession directly and pass the session type and seconds or users can keep calling the start methods for respective sessionTypes. It's also possible to start a session for a custom seconds, in a middle of a session as an example.